feat: add menu(beta) and music(beta)
