+++ package.json
@@ -10,7 +10,10 @@
     "pixi.js": "^5.3.7",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.2",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
