+++ src/components/menu/menu.module.scss
@@ -1,0 +1,35 @@
+//@import '../sass/shadow-hover.scss';
+
+.menu-hidden {
+  transition: all 0.6s;
+  position: fixed;
+  top: -66vh;
+  left: 25vw;
+}
+
+.menu_shown {
+  transition: all 0.6s;
+  position: fixed;
+  top: 29vh;
+  left: 35vw;
+}
+
+.menu {
+  z-index: 1000;
+  background-color: rgba(239, 241, 234, 0.925);
+  width: 50vw;
+  height: 30vh;
+  display: flex;
+  justify-content: space-around;
+  font-size: 2rem;
+  align-items: center;
+  flex-direction: column;
+  border: 1px solid black;
+  transition: all 0.3s;
+
+  div {
+    transition: all 0.3s;
+    cursor: pointer;
+    //@include shadow-hover;
+  }
+}
